+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- `Year`, `Month`, `Date`, `Time`, `LocalDateTime` and `ZonedDateTime` are comparable from now on.
+  Next methods had added to these types:
+    - `compare()`
+    - `greaterThan()`
+    - `greaterThanOrEqual()`
+    - `lessThan()`
+    - `lessThanOrEqual()`
+
 ## v0.1.1 - 2016-10-10
 ### Added
 - Added shortcut factory `ZonedDateTime::nowUtc()` which returns current time in the UTC timezone.